Changed Player Files + Tests
